use global state and create MovieCards component
