+++ client/src/pages/SearchMovies.js
@@ -5,13 +5,12 @@ import { searchTMDB } from '../utils/API';
 import { cleanMovieData } from '../utils/movieData';
 
 // import react-bootstrap components
-import { Form, Button, Container, CardColumns, Jumbotron } from 'react-bootstrap';
+import { Form, Button, Container, Jumbotron } from 'react-bootstrap';
 
 // import custom components
-import MovieCard from '../components/MovieCard'
+import MovieCards from '../components/MovieCards'
 
 const SearchMovies = () => {
-    // local state for search
     const [searchInput, setSearchInput] = useState('');
     const [searchedMovies, setSearchedMovies] = useState([]);
 
@@ -63,20 +62,9 @@ const SearchMovies = () => {
 
             <Container>
                 <h2 className="results-heading">
-                    {searchedMovies.length > 0 &&
-                    `Viewing ${searchedMovies.length} results:`}
+                    {searchedMovies.length > 0 && `Viewing ${searchedMovies.length} results:`}
                 </h2>
-                <CardColumns>
-                    {searchedMovies.map((movie) => {
-                        return (
-                            <MovieCard
-                                displayTrailer
-                                movie={movie}
-                                saveBtn
-                                deleteBtn />
-                        )
-                    })}
-                </CardColumns>
+                <MovieCards displayTrailers='true' moviesToDisplay={searchedMovies} />
             </Container>
         </>
     );
